moving styles to dedicated directories within components
