+++ anki/src/components/styles/Deck.css
@@ -5,7 +5,7 @@
     transform: translate(-50%, 0);
     border-top-left-radius: 20px;
     border-top-right-radius: 20px;
-    transition: width 100ms, height 100ms,
+    transition: width 100ms, height 100ms;
 }
 
 .deck-wrapper {
@@ -18,8 +18,8 @@
 
 .button-wrapper {
     position: relative;
-    padding: 0px;
-    border: 0px;
+    padding: 0;
+    border: 0;
     background: none;
     transition: width 100ms, height 100ms;
 }
@@ -27,7 +27,7 @@
 .deck {
     position: absolute;
     width: 100%;
-    bottom: 0px;
+    bottom: 0;
     border-radius: 4px;
     display: flex;
     justify-content: center;
