+++ anki/src/components/styles/Header.css
@@ -1,6 +1,6 @@
 .main-header {
     position: absolute;
-    top: 0px;
+    top: 0;
     display: flex;
     height: 40px;
     width: 100%;
@@ -14,5 +14,5 @@
     position: absolute;
     height: 100%;
     width: 120px;
-    right: 0px;
+    right: 0;
 }
