+++ anki/src/components/styles/StatusBar.css
