+++ anki/src/features/deckspace/components/DeckSpace.tsx
@@ -5,7 +5,7 @@ import { Deck, StatusBar, FadeInOut } from "components";
 import { getMe } from 'api';
 import { getDecks } from "features/deckspace/api";
 import { DeckType } from "types";
-import "./DeckSpace.css";
+import "./styles/DeckSpace.css";
 
 const DeckSpace = () => {
     const { username, deckname } = useParams();
