+++ anki/src/features/deckspace/components/styles/DeckInfo.css
@@ -114,11 +114,11 @@
 }
 
 .play-button {
-    background-image: url('../assets/play-button.svg');
+    background-image: url('../../assets/play-button.svg');
 }
 
 .edit-button {
-    background-image: url('../assets/edit-button.svg');
+    background-image: url('../../assets/edit-button.svg');
 }
 
 .deckinfo-closer a {
