+++ anki/src/features/editmode/components/SaveChangesBar.tsx
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 import { ThemeContext } from "context";
 import { ButtonSwitch } from "components";
-import './SaveChangesBar.css';
+import './styles/SaveChangesBar.css';
 
 type Props = {
     saveFunction: () => void,
