+++ anki/src/features/trainmode/components/styles/ProgressBar.css
