+++ anki/src/features/trainmode/components/styles/TrainMode.css
